requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
python-whois==0.9.4
python-multipart==0.0.6
